Added for add and edit user

+++ src/main/java/com/ims/repository/UserDetailsRepository.java
@@ -6,7 +6,7 @@ import com.ims.model.User;
 
 
 
-public interface UserDetailsRepository extends JpaRepository<User, Long> {
+public interface UserDetailsRepository extends JpaRepository<User, Integer> {
 
 	User findByUserName(String userName);
